Notifications: move fake request to methods.utils
