+++ fir_notifications/methods/__init__.py
@@ -1,19 +1,6 @@
 import json
 
 from django.template import Template, Context
-from django.conf import settings
-
-
-class FakeRequest(object):
-    def __init__(self):
-        self.base = settings.EXTERNAL_URL
-        if self.base.endswith('/'):
-            self.base = self.base[:-1]
-
-    def build_absolute_uri(self, location):
-        return "{}{}".format(self.base, location)
-
-request = FakeRequest()
 
 
 class NotificationMethod(object):
